// ui-parts/carousel/src/pagination/shape/shape.interfaces.ts
export interface PaginationShapeStyles {
  size: number | string
  margin?: number | string
  borderWidth?: number | string
  borderRadius?: number | string
  activeSize?: number | string
  activeMargin?: number | string
  activeBorderWidth?: number | string
  activeBorderRadius?: number | string
}